Primefaces-Hibernate-Spring-Integration-Sample
Primefaces-Hibernate-Spring-Integration-Demo
This Demo is a small brief of using JSF With Spring and Hibernate..
The main page is http://localhost:8080/Primefaces-Hibernate-Spring-Integration-Sample
App Story : This app is meeting organization app.
You can organize a meeting with all members, give roles to members, dedicate times for this meetings, and don't worry about time intervals.
explore and enjoy ..! :smile: